+++ experimental/touchflow-cpp/src/settings_app.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+namespace touchflow {
+
+int run_settings_app(int argc, char** argv);
+
+}  // namespace touchflow