+++ experimental/touchflow-cpp/src/settings_main.cpp
@@ -1,0 +1,5 @@
+#include "settings_app.hpp"
+
+int main(int argc, char** argv) {
+    return touchflow::run_settings_app(argc, argv);
+}